refactor: move BundleCommonBase into dao package
BundleCommonBase's only subclass, MultiTenantDBShardingBundleBase, already
lives in io.appform.dropwizard.sharding.dao. Moving BundleCommonBase there
too puts DBShardingBundleBase in the same package as the class that declares
getBucketInfo(), so the protected method is now directly accessible via
ordinary same-package visibility. This removes the need for the trivial
pass-through override in MultiTenantDBShardingBundleBase that previously
existed only to bridge visibility across the old package boundary.
